Proper SGI aspect ratio (fixes issue #9) + GLES2 rasterize by default + resizable framebuffers
Display / present path:
- Zero-readback present: display quad samples the gles2 front FBO texture directly; glReadPixels remains only for GEN_FRAME_PPM_FILES dumps
- All display textures are exact-size non-power-of-two (NPOT) (y-flip moved to positive texcoord space, dropping the GL_REPEAT/POT requirement and texSize uniform)
- SGI pixel aspect ratio (PAR): SGI pixels were 16:15 wide (5:4 grid on 4:3 CRT glass), so the present pass stretches 16/15 horizontally over a framebuffer carrying 15/16 the window's columns -- fills the window exactly, circles are slightly wide just like on the real hardware. SGI_PAR=0 native / ?par=0 web disables.

Resizable, window-sized framebuffers:
- Framebuffer initializes to and tracks the window's logical size; new rasterizer_resize() in the dispatch API (ref reallocs heap buffers, gles2 rebuilds FBOs atomically); demos notified via REDRAW (IRIS GL style)
- screen_vertex x/y widened uint16 -> int32; size capped only by GL_MAX_TEXTURE_SIZE, queried at context creation
- Implied reshape: gl.c records perspective/window/ortho2 calls and re-derives the projection at the new aspect on resize -- the GLUT reshape pattern, done by the shim so demo code stays untouched. Transforms the demo composed onto the projection stack (buttonfly and bounce translate after perspective in MPROJECTION mode) are preserved: the reapply replaces only the recorded projection part (E x P_new), so buttonfly's flip fills the screen exactly at any window size
- keepaspect() implemented with IRIX timing: pre-winopen applies at window creation; post-winopen stays pending until winconstraints() (buttonfly's SPACE toggle). Declared-aspect demos (jello, buttonfly, logo, ideas, cedit) get conforming native windows / centered web framebuffers; no visible letterboxing in either tier
- REDRAW is delivered on framebuffer changes unconditionally -- IRIS GL queued it by default with no qdevice() needed (fixes twilight going black on resize)
- Per-demo shim quirks (rule #1: demo sources stay untouched): arena keeps its hardcoded 800:480 projection aspect; flight gets a classic fixed-size 800x480 framebuffer scaled to the window (its panel viewports are compile-time constants)
- Window geometry: logical (mouse) and drawable (GL) sizes tracked separately with a float pixelScale, refreshed on every resize (fixes the web fullwindow-canvas stretch regression)

Also:
- gles2 rasterizer is now the default (GLES2_RASTERIZER=ref for CPU reference)
- Makefiles: headers are real object prerequisites, so header layout changes rebuild all dependents (prevents mixed-ABI builds)
- Web builds: -sALLOW_MEMORY_GROWTH for runtime-sized framebuffer allocation
- tests/native-resize: harness driving real SDL window resizes, asserting conforming windows and framebuffer re-fit
- Smoke baselines regenerated (buttonfly button now at authentic IRIX proportions; arena/flight letterboxing changed)
- Demo name is stamped into each binary at build time (makefiles/gl_appname.c, APPNAME from the Makefile); winopen ignores the passed title, fixing web titles ("this.program") and letting per-demo quirks match by name on both tiers

+++ libs/libgl/basic_types.h
@@ -25,7 +25,9 @@ typedef uint8_t bool;
 
 typedef struct screen_vertex
 {
-    uint16_t x, y;
+    int32_t x, y;   // pixels in 1/SCREEN_VERTEX_V2_SCALE fixed point; signed
+                    // so coordinate differences are safe, 32-bit so any
+                    // realistic framebuffer size fits (2^31/32 = 67M pixels)
     uint32_t z;
     uint8_t r, g, b, a;
 } screen_vertex;
